Change DeletedAtUnixTimestamp type from int to long
Updated the DeletedAtUnixTimestamp property in User.cs to use long instead of int. This ensures support for larger Unix millisecond timestamps and prevents potential overflows for dates far from the epoch.

+++ Sources/Mattermost/Models/Users/User.cs
@@ -46,7 +46,7 @@ namespace Mattermost.Models.Users
         /// The time in milliseconds a user was deleted.
         /// </summary>
         [JsonPropertyName("delete_at")]
-        public int DeletedAtUnixTimestamp { get; set; }
+        public long DeletedAtUnixTimestamp { get; set; }
 
         /// <summary>
         /// The <see cref="DateTime"/> a user was deleted in UTC format.
